# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API untuk platform ujian online IPPAT (Fiber + GORM + PostgreSQL). Semua response dibungkus format {success, message, data, errors}.\nEndpoint yang butuh login (guru/admin/peserta) ditandai kunci gembok di UI ini — klik \"Authorize\" dan isi `Bearer \u003ctoken\u003e` yang didapat dari /api/auth/login atau /api/auth/register.",
        "title": "IPPAT Backend API",
        "contact": {
            "name": "IPPAT Backend Team"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/jadwal": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jadwal"
                ],
                "summary": "List jadwal ujian",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Nomor halaman",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Jumlah data per halaman",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_jadwal_dto.JadwalListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jadwal"
                ],
                "summary": "Buat jadwal ujian baru",
                "parameters": [
                    {
                        "description": "Data jadwal",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_internal_modules_jadwal_dto.CreateJadwalRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_jadwal_dto.JadwalResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
...
        "/jadwal/aktif/hari-ini": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil daftar jadwal ujian yang berlangsung hari ini untuk kelas peserta yang sedang login, lengkap dengan status pengerjaannya (belum mulai/sedang ujian/selesai).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jadwal"
                ],
                "summary": "Jadwal ujian aktif hari ini untuk peserta yang login",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/backend_internal_modules_jadwal_dto.JadwalAktifResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
        "/jadwal/bank-soal/{bank_soal_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jadwal"
                ],
                "summary": "List jadwal berdasarkan bank soal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Bank Soal",
                        "name": "bank_soal_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Nomor halaman",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Jumlah data per halaman",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_jadwal_dto.JadwalListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
        "/jadwal/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jadwal"
                ],
                "summary": "Detail jadwal ujian",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Jadwal",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_jadwal_dto.JadwalResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jadwal"
                ],
                "summary": "Update jadwal ujian",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Jadwal",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data jadwal",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_internal_modules_jadwal_dto.UpdateJadwalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_jadwal_dto.JadwalResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jadwal"
                ],
                "summary": "Hapus (soft delete) jadwal ujian",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Jadwal",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
        "/jadwal/{id}/restore": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jadwal"
                ],
                "summary": "Pulihkan jadwal yang sudah dihapus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Jadwal",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "backend_internal_helpers.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "backend_internal_modules_jadwal_dto.CreateJadwalRequest": {
            "type": "object",
            "required": [
                "durasi",
                "id_bank_soal",
                "id_kelas",
                "nama_ujian",
                "tingkat",
                "wkt_mulai",
                "wkt_selesai"
            ],
            "properties": {
                "acak_opsi": {
                    "type": "integer"
                },
                "acak_soal": {
                    "type": "integer"
                },
                "durasi": {
                    "type": "integer",
                    "minimum": 1
                },
                "id_bank_soal": {
                    "type": "string"
                },
                "id_kelas": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "nama_ujian": {
                    "type": "string"
                },
                "tingkat": {
                    "type": "string"
                },
                "wkt_mulai": {
                    "type": "string"
                },
                "wkt_selesai": {
                    "type": "string"
                }
            }
        },
        "backend_internal_modules_jadwal_dto.JadwalAktifResponse": {
            "type": "object",
            "properties": {
                "acak_opsi": {
                    "type": "integer"
                },
                "acak_soal": {
                    "type": "integer"
                },
                "durasi": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "id_bank_soal": {
                    "type": "string"
                },
                "id_nilai": {
                    "type": "string"
                },
                "nama_bank_soal": {
                    "type": "string"
                },
                "nama_ujian": {
                    "type": "string"
                },
                "status_pengerjaan": {
                    "type": "string"
                },
                "tingkat": {
                    "type": "string"
                },
                "wkt_mulai": {
                    "type": "string"
                },
                "wkt_selesai": {
                    "type": "string"
                }
            }
        },
        "backend_internal_modules_jadwal_dto.JadwalListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_modules_jadwal_dto.JadwalResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "backend_internal_modules_jadwal_dto.JadwalResponse": {
            "type": "object",
            "properties": {
                "acak_opsi": {
                    "type": "integer"
                },
                "acak_soal": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "durasi": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "id_bank_soal": {
                    "type": "string"
                },
                "id_jurusan": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_modules_jadwal_dto.JurusanItem"
                    }
                },
                "id_kelas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_modules_jadwal_dto.KelasItem"
                    }
                },
                "nama_bank_soal": {
                    "type": "string"
                },
                "nama_ujian": {
                    "type": "string"
                },
                "tingkat": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "wkt_mulai": {
                    "type": "string"
                },
                "wkt_selesai": {
                    "type": "string"
                }
            }
        },
        "backend_internal_modules_jadwal_dto.JurusanItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "id_jurusan": {
                    "type": "string"
                },
                "nama_jurusan": {
                    "type": "string"
                }
            }
        },
        "backend_internal_modules_jadwal_dto.KelasItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "id_kelas": {
                    "type": "string"
                },
                "nama_kelas": {
                    "type": "string"
                }
            }
        },
        "backend_internal_modules_jadwal_dto.UpdateJadwalRequest": {
            "type": "object",
            "required": [
                "durasi",
                "id_bank_soal",
                "id_kelas",
                "nama_ujian",
                "tingkat",
                "wkt_mulai",
                "wkt_selesai"
            ],
            "properties": {
                "acak_opsi": {
                    "type": "integer"
                },
                "acak_soal": {
                    "type": "integer"
                },
                "durasi": {
                    "type": "integer",
                    "minimum": 1
                },
                "id_bank_soal": {
                    "type": "string"
                },
                "id_kelas": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "nama_ujian": {
                    "type": "string"
                },
                "tingkat": {
                    "type": "string"
                },
                "wkt_mulai": {
                    "type": "string"
                },
                "wkt_selesai": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Ketik \"Bearer\" diikuti spasi dan JWT token. Contoh: \"Bearer eyJhbGciOi...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}